+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [standard] GRPC default functions for ListenAndServe and Gateway
-
 ### Changed
 
 ### Deprecated
@@ -20,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v1.78.0] - 2023-06-14
+
+### Added
+
+- [standard] GRPC default functions for ListenAndServe and Gateway
 
 ## [v1.77.0] - 2023-05-12
 
@@ -942,7 +946,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.77.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.78.0...HEAD
+[v1.78.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.77.0...v1.78.0
 [v1.77.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.76.0...v1.77.0
 [v1.76.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.75.0...v1.76.0
 [v1.75.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.74.0...v1.75.0